Move DemoFragment to separate module; rename to HomeFragment
