+++ app/build.gradle
@@ -48,6 +48,7 @@ apply plugin: 'kotlin-android'
 
 dependencies {
     implementation project(path: ':base')
+    implementation project(path: ':home')
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation project(path: ':user')
