+++ app/src/main/res/navigation/nav_graph.xml
@@ -11,12 +11,15 @@
         >
         <action
             android:id="@+id/actionLoginFragmentToDemoFragment"
-            app:destination="@id/demoFragment" />
+            app:destination="@id/demoFragment"
+            app:popUpTo="@+id/nav_graph"
+            app:popUpToInclusive="true"
+            />
     </fragment>
 
     <fragment
         android:id="@+id/demoFragment"
-        android:name="com.example.appName.demoFeature.DemoFragment"/>
+        android:name="com.example.home.HomeFragment"/>
 
 
 </navigation>
