+++ home/src/main/java/com/example/home/HomeFragment.kt
@@ -1,15 +1,14 @@
-package com.example.appName.demoFeature
+package com.example.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.appName.R
 
-class DemoFragment : Fragment() {
+class HomeFragment : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_demo, container, false)
+        return inflater.inflate(R.layout.fragment_home, container, false)
     }
 }
