+++ settings.gradle
@@ -1,3 +1,4 @@
+include ':home'
 include ':user'
 include ':base'
 include ':app'
